Created new faulttolerance_repeat_tests project and moved RepeatFaultTolerance.java
