+++ dev/com.ibm.ws.microprofile.faulttolerance_repeat_tests/.settings/bndtools.core.prefs
@@ -1,0 +1,2 @@
+compileErrorAction=build
+eclipse.preferences.version=1
