+++ dev/com.ibm.ws.microprofile.faulttolerance_repeat_tests/bnd.bnd
@@ -1,0 +1,29 @@
+#*******************************************************************************
+# Copyright (c) 2020 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#
+# Contributors:
+#     IBM Corporation - initial API and implementation
+#*******************************************************************************
+-include= ~../cnf/resources/bnd/bundle.props
+
+bVersion=1.0
+	
+Bundle-Name: MP Fault Tolerance Repeat FAT test infrastructure
+Bundle-SymbolicName: com.ibm.ws.microprofile.faulttolerance_repeat_tests
+Bundle-Description: MP Fault Tolerance Repeat FAT test infrastructure; version=${bVersion}
+
+Export-Package: com.ibm.ws.microprofile.faulttolerance.fat.repeat
+
+javac.source: 1.8
+javac.target: 1.8
+
+test.project: true
+
+-buildpath: \
+        com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
+        fattest.simplicity
+
